pyproject.toml
[project]
name = "yabench"
version = "0.1.0"
description = "Yet Another Benchmark for LLM APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
yabench = "yabench.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["yabench", "datasets"]
